Added the patch parents functionality. This is specifically for the database portion. 1. Changed functionality to simply add a parent. 2. Can copy the Measurements from parent. 3. Can copy the Additional Data from parent.
